refactor: move logic to retrieve data for blog to container
