+++ src/pages/blog.tsx
@@ -4,7 +4,6 @@ import Head from '../components/Head';
 import Layout from '../components/Layout';
 import Stars from '../components/Stars';
 import Content from '../components/Content';
-import {Work} from '../types';
 import Blog from '../components/Blog';
 
 interface BlogPageProps {
@@ -13,9 +12,6 @@ interface BlogPageProps {
 
 const BlogPage: React.FC<BlogPageProps> = ({data}) => {
   const siteTitle = data.site.siteMetadata.title;
-  const posts: Work[] = data.allMarkdownRemark.edges.filter(
-    work => work.node.frontmatter.type === 'post'
-  );
 
   return (
     <Layout title={siteTitle}>
@@ -36,7 +32,7 @@ const BlogPage: React.FC<BlogPageProps> = ({data}) => {
         addEventListeners={false}
       />
       <Content>
-        <Blog posts={posts} />
+        <Blog />
       </Content>
     </Layout>
   );
@@ -48,26 +44,6 @@ interface PageQueryData {
       title: string;
     };
   };
-  allMarkdownRemark: {
-    edges: {
-      node: {
-        id: string;
-        excerpt: string;
-        fields: {
-          slug: string;
-        };
-        frontmatter: {
-          date: string;
-          title: string;
-          description: string;
-          tags: string[];
-          type: string;
-          url: string;
-          featuredImage: any;
-        };
-      };
-    }[];
-  };
 }
 
 export const pageQuery = graphql`
@@ -75,35 +51,6 @@ export const pageQuery = graphql`
     site {
       siteMetadata {
         title
-      }
-    }
-    allMarkdownRemark(
-      filter: {frontmatter: {published: {ne: false}}}
-      sort: {fields: [frontmatter___date], order: DESC}
-    ) {
-      edges {
-        node {
-          id
-          excerpt
-          fields {
-            slug
-          }
-          frontmatter {
-            date(formatString: "MMM DD, YYYY")
-            title
-            description
-            tags
-            type
-            url
-            featuredImage {
-              childImageSharp {
-                fluid(maxWidth: 800) {
-                  ...GatsbyImageSharpFluid
-                }
-              }
-            }
-          }
-        }
       }
     }
   }
